refactor: update chat components imports to use shadcn directory
