+++ features/chat/components/chat-message-item.tsx
@@ -5,13 +5,13 @@ import Image from 'next/image';
 import { format } from 'date-fns';
 import { CornerUpLeft, Pencil, Trash2, Smile } from 'lucide-react';
 
-import { Button } from '@/components/ui/button';
+import { Button } from '@/components/shadcn/button';
 import {
   Tooltip,
   TooltipContent,
   TooltipProvider,
   TooltipTrigger,
-} from '@/components/ui/tooltip';
+} from '@/components/shadcn/tooltip';
 import { ChatMessage } from '@/types/chat';
 import {
   parseMentions,
