+++ features/chat/components/chat-sidebar.tsx
@@ -16,9 +16,9 @@ import {
   ChevronRight,
   Loader2,
 } from 'lucide-react';
-import { Input } from '@/components/ui/input';
-import { Button } from '@/components/ui/button';
-import { Badge } from '@/components/ui/badge';
+import { Input } from '@/components/shadcn/input';
+import { Button } from '@/components/shadcn/button';
+import { Badge } from '@/components/shadcn/badge';
 
 import {
   Sheet,
@@ -27,7 +27,7 @@ import {
   SheetHeader,
   SheetTitle,
   SheetTrigger,
-} from '@/components/ui/sheet';
+} from '@/components/shadcn/sheet';
 import { useChatRooms } from '@/hooks/chat/use-chat-rooms';
 import { useEmployees } from '@/hooks/employee';
 import { useCreateDirectRoom } from '@/hooks/chat/use-chat-mutations';
